Allow Verifier to derive session key prior to verifying.
The prior interface only allowed accessing the session key after
verifying it. However as part of verifying session encryption works,
the Verifier may receive the user session verifier from the user
encrypted using the session key. In this case the session verifier
must first be decrypted with the session key prior to verifying it.
Thus in this special case the session key is needed prior to verification.